requirements.txt
python-can>=4.0.0
paho-mqtt>=1.6.0
influxdb-client>=1.36.0
PyQt5>=5.15.0
